more cleanup of the tests
